-Termine la creacion de cuenta con las validaciones de usuario, empresa y cliente. Faltaría hashear el password pero lo agregue como otra tarea en el trello. -Termine la busqueda de clientes en el ABM de clientes.

+++ app/FrbaCommerce/Modelos/Usuario.cs
@@ -7,7 +7,7 @@ namespace FrbaCommerce.Modelos
 {
     class Usuario
     {
-        public Int32 id { get; set; }
+        public Int64 id { get; set; }
         public Int32 login_fallidos { get; set; }
         public String username { get; set; }
         public String password { get; set; }
